Fix user dn when adding groups members via ldap
The base_user_dn in load_test.py already contains the 'uid={}' string.
As a consequence of this mistake, the "member" attribute of the group
was populated with the wrong string for user dns and the "memberof"
attribute of the users were not populated.

This commit fixes the mod_group_users_ldap method to use base_user_dn
as a template for formating rather than a suffix for appending to the
uid.